refactor: 레파지토리 인터페이스 추가
